Put tests in separate files.
Makes it easier to disable them when testing specific subsets of the package.
