+++ test/util.jl
@@ -1,0 +1,13 @@
+macro grab_output(ex)
+    quote
+        mktemp() do fname, fout
+            ret = nothing
+            open(fname, "w") do fout
+                redirect_stdout(fout) do
+                    ret = $(esc(ex))
+                end
+            end
+            ret, read(fname, String)
+        end
+    end
+end
